% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyInterception.R
\name{canopyInterception}
\alias{canopyInterception}
\title{Snow canopy interception model (Hedstrom and Pomeroy, 1998)}
\usage{
canopyInterception(Tair, LAI, P, Smean = 6.6, L0 = 0)
}
\arguments{
\item{Tair}{(deg C) air temperature}

\item{LAI}{()      Leaf area index}

\item{P}{(mm SWE) snow precipitation}

\item{Smean}{(kg/m^2)  snow loading coefficient}

\item{L0}{(mm SWE) initial snow load}

\item{c}{() dimensionless unloading coefficient = 0.678}
}
\value{
Canopy interception (mm)
}
\description{
THis function computer evapotranspiration based on radiation, conductance etc
}
\author{
Bar Avni
}

