% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoffCalc.R
\name{runoffCalc}
\alias{runoffCalc}
\title{Snow canopy interception model (Hedstrom and Pomeroy, 1998)}
\usage{
runoffCalc(P, I, InfilCap, MaxStorage, ET)
}
\arguments{
\item{P}{(mm SWE) snow precipitation}

\item{I}{(mm) Canopy interception}

\item{InfilCap}{(mm/day) constant infiltration capacity}

\item{MaxStorage}{(mm) maximum capacity of ground storage}

\item{ET}{(mm/day) evapotranspiration}
}
\value{
daily runoff depth   (mm)
}
\description{
THis function computer evapotranspiration based on radiation, conductance etc
}
\author{
Bar Avni
}

