% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thornthwaite.R
\name{Thornthwaite}
\alias{Thornthwaite}
\title{THornwaite Equation}
\usage{
Thornthwaite(day, year, Tair, TMDH)
}
\arguments{
\item{year}{}

\item{Tair}{(deg C) air temperature}

\item{TMDH}{Total monthly daylight hours}

\item{nday}{}
}
\value{
monthly potential ET
}
\description{
THis function computer evapotranspiration
}
\author{
Bar Avni
}

